move HamError
