+++ include/public/util/HamError.hpp
